Fix skins, disconnect packet, bump protocol version to 8

+++ src/main/java/org/itxtech/nemisys/network/protocol/spp/PlayerLoginPacket.java
@@ -26,7 +26,7 @@ public class PlayerLoginPacket extends SynapseDataPacket {
         this.putString(this.address);
         this.putInt(this.port);
         this.putByte(this.isFirstTime ? (byte) 1 : (byte) 0);
-        this.putShort(this.cachedLoginPacket.length);
+        this.putInt(this.cachedLoginPacket.length);
         this.put(this.cachedLoginPacket);
     }
 
@@ -36,6 +36,6 @@ public class PlayerLoginPacket extends SynapseDataPacket {
         this.address = this.getString();
         this.port = this.getInt();
         this.isFirstTime = this.getByte() == 1;
-        this.cachedLoginPacket = this.get(this.getShort());
+        this.cachedLoginPacket = this.get(this.getInt());
     }
 }
